Consolidate Session modules
The following is an excerpt from the discussion at
https://github.com/binarylogic/authlogic/pull/642

```
**I have two goals:**

1. **Clarify how existing mixins interact** - Currently, it is difficult to trace
  - the execution of "super chains" like `#credentials=`, and
  - the order of callbacks (eg. the four `persist` callbacks).
  - Consolidating the mixins into a single file solves this goal. See
    [observations](https://github.com/binarylogic/authlogic/pull/642#issuecomment-445440228)
    above.
2. **Extract new _encapsulated_ classes**
  - Currently, there is no encapsulation. All 2 kLoC are mixed into the same
    `Base`, sharing the same instance variables, methods, etc. For example,
    `attempted_record` is used in nine mixins. `klass` is used in twenty. It
    seems as if these mixins are separate things, but in reality they have
    complicated dependencies on each other. They are *modules*, but they're not
    *independent*.
  - Consolidating the mixins is an *intermediate step* that helps us identify
    and extract new, *encapsulated* classes like those I've suggested above
    ([here](https://github.com/binarylogic/authlogic/pull/642#issue-235567497)
    and
    [here](https://github.com/binarylogic/authlogic/pull/642#issuecomment-445440228)).
    I could refine those suggestions a bit, if that would help. If anyone sees a
    class they'd extract, I'd love to hear other ideas.
```
