+++ .rubocop.yml
@@ -49,6 +49,11 @@ Metrics/AbcSize:
     # crypto_providers/wordpress is deprecated so we will not attempt to
     # improve its quality.
     - lib/authlogic/crypto_providers/wordpress.rb
+    # After consolidating the 23 session modules into one file (see discussion
+    # in https://github.com/binarylogic/authlogic/pull/642) some methods are
+    # too big. This is a temporary exclusion. This would be an entry in
+    # `.rubocop_todo.yml` if rubocop allowed `Exclude`s in both simultaneously.
+    - 'lib/authlogic/session/base.rb'
     # In an ideal world tests would be held to the same ABC metric as production
     # code. In practice, time spent doing so is not nearly as valuable as
     # spending the same time improving production code.
@@ -67,6 +72,11 @@ Metrics/CyclomaticComplexity:
     # crypto_providers/wordpress is deprecated so we will not attempt to
     # improve its quality.
     - lib/authlogic/crypto_providers/wordpress.rb
+    # After consolidating the 23 session modules into one file (see discussion
+    # in https://github.com/binarylogic/authlogic/pull/642) some methods are
+    # too big. This is a temporary exclusion. This would be an entry in
+    # `.rubocop_todo.yml` if rubocop allowed `Exclude`s in both simultaneously.
+    - 'lib/authlogic/session/base.rb'
 
 # Aim for 80, but 100 is OK.
 Metrics/LineLength:
