+++ lib/authlogic.rb
@@ -32,29 +32,7 @@ path = File.dirname(__FILE__) + "/authlogic/"
   "acts_as_authentic/single_access_token",
   "acts_as_authentic/base",
 
-  "session/activation",
-  "session/active_record_trickery",
-  "session/brute_force_protection",
-  "session/callbacks",
-  "session/cookies",
-  "session/existence",
-  "session/foundation",
-  "session/http_auth",
-  "session/id",
-  "session/klass",
   "session/magic_column/assigns_last_request_at",
-  "session/magic_columns",
-  "session/magic_states",
-  "session/params",
-  "session/password",
-  "session/perishable_token",
-  "session/persistence",
-  "session/priority_record",
-  "session/scopes",
-  "session/session",
-  "session/timeout",
-  "session/unauthorized_record",
-  "session/validation",
   "session/base"
 ].each do |library|
   require path + library
